Components Done; Context Next
Completed the skeletal structure of the components, and moved them into their own folders.

Next would be ensuring that the forms actually update state.

After having the ability to update state I will then consolidate, and clean up with context.

Focused on Budget component because I thought it would be easy but turns out no. Tomorrow will focus on the other components, and return back to it when I can