Connected Server
Connected the endpoints to the react client.

Things to note:

weddings.budget is not accessible as it is reading the value as unkown. It may have to do with PropType.

Post/Patch requests still not functioning. Same error message is received when making a post request.

Receieve an error message when using the delete functions